pyproject.toml
[project]
name = "rag-youtube"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "faiss-cpu>=1.10.0",
    "langchain>=0.3.24",
    "langchain-community>=0.3.22",
    "langchain-core>=0.3.55",
    "langchain-experimental>=0.3.4",
    "langchain-openai>=0.3.14",
    "langchain-text-splitters>=0.3.8",
    "python-dotenv>=1.1.0",
    "streamlit>=1.44.1",
    "tiktoken>=0.9.0",
    "youtube-transcript-api>=1.0.3",
]
